refactor(server): rename cycle_time to cycle_time_ms and upgrade type to u32
- Rename `cycle_time` field to `cycle_time_ms` in configuration and triggers.
- Change data type from `u16` to `u32` to allow for longer cycle durations.
- Update `config.toml` and sample configuration to reflect the naming change.

+++ rt-server-rs/src/cycle/interval.rs
@@ -13,11 +13,11 @@ use std::{thread, time};
 use super::CycleTrigger;
 
 pub struct IntervalTrigger {
-    cycle_ms: u16,
+    cycle_ms: u32,
 }
 
 impl IntervalTrigger {
-    pub fn new(cycle_ms: u16) -> Self {
+    pub fn new(cycle_ms: u32) -> Self {
         IntervalTrigger { cycle_ms }
     }
 }
